refactor(validation): 이름, 링크, 드롭다운, 설명에 대한 validation 분리 및 에러 상수 분리
